+++ src/constants.ts
@@ -143,6 +143,7 @@ export const PACKAGES_TO_GET_INFO_FOR = [
   'playwright',
   'lodash',
   'lodash-es',
+  'qunit',
   ...CHECKED_LODASH_METHODS.map((method) => `lodash.${method}` as const),
 
   'astro',